+++ src/app/(private)/gastos/[id]/page.client.jsx
@@ -64,6 +64,7 @@ const ExpensePageClient = ({ expense }) => {
         color: COLORS.GREEN,
         onClick: () => push(PAGES.EXPENSES.CLONE(expense.id)),
         text: 'Clonar',
+        collapsedTooltip: 'Clonar gasto',
       });
     }
 
@@ -77,6 +78,7 @@ const ExpensePageClient = ({ expense }) => {
         basic: true,
         onClick: handleAnular,
         text: "Anular",
+        collapsedTooltip: "Anular gasto",
         width: CONTENT_SIZES.FIT,
       });
     }